[Feature(drive.c)] Implement application level drive interface
If included, this commit implements the drive.c implementation file that contains drive abstraction layer of the TB6612FNG motor driver. This file allows handles the initialization of the motor driver as well as stopping and setting the motor speeds and modes. The modes include forward, reverse, as well as sharp, mid, and wide arcturns with varying speeds.