[Feature(assert_handler.c)] Implement assert handling functionality
If included, commit implements the assert handling structure to be used throughout the project. The assert implementation is to be used to ensure that peripherals and io pins are initialized properly before use as well as numerous other safety checks. The implementation makes use of usart polling to inform the reader of the location of the assert via program counter address. With the address the user may retrieve the line number and file name. The assert also blinks an LED to indicate to the user that an assert has been triggered.