[Feature(adc.c/dma.c)] Implement ADC conversion with DMA for data transfer
If included, commit will implement the ADC driver functionality that enables and configures the ADC peripheral in SCAN mode with continuous conversion that scans the 4 channels to be used for the light sensors. DMA implementation file enables and configures the DMA peripheral to be used for data transfer from ADC peripheral to a buffer in memory from which the user can retrieve digital values from the light sensor.